created new test.js files for get-product and product api, and imported those files into tests.js
