+++ test/product.api.test.js
@@ -1,0 +1,17 @@
+const test = QUnit.test;
+
+QUnit.module('product api');
+
+test('creates product api test', (assert) => {
+    // Arrange
+    // Set up your paramenters and expectations
+    const product = {
+
+    };
+
+    // Act
+    // Call the function you're testing and set the result to a const
+
+    // Assert
+    assert.deepEqual(result, product);
+});
